divide fetching logics at stage page
